Renaming module name to preserve pep8 conventions: removing '_'.
