+++ comparison/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ['ModelComparison']
 
-from .algorithm_selection import ModelComparison
+from .modelcomparison import ModelComparison
